build: add --enable-pmix-bootstrap support
Problem: on some platforms PMIx is the preferred mechanism
to use for bootstrapping Flux.

Prepare to add PMIx support to the Flux broker by adding
an "opt in" configure option: --enable-pmix-bootstrap.
If specified, pkg-config is used to locate a suitable pmix
package.  Configure fails if pmix is requested but not found.

Add PMIX_LIBS and PMIX_CFLAGS to the broker Makefile.am.

Co-authored-by: Jim Garlick <garlick.jim@gmail.com>