broker/modservice : add FRIPP metric hook
Add a hook to the built-in module services to optionally
send FRIPP message count metrics on every reactor loop pass.

To enable:
  flux module debug modulename --setbit=0x1000

In the debug RPC handler, activate the prepare watcher
if this bit is set.

In the prepare watcher, if the bit is set, send metric
(a flux_log call is currently a stand-in for the metric).
Otherwise disable the prepare watcher.

Logic was added so the prepare watcher does not get destroyed
on first use, and to ensure the RUNNING keepalive message is
only sent to the broker once.