flux-start: add --list option
Problem: it is easy to lose track of local test instances.

Have flux-start create all test instances for a given user in
a directory /tmp/flux-userid-N.

Add a flux start --list option that lists the URIs for all local
test instances by scanning this directory for local-0 files.
List URIs corresponding to found local-0 files, and also connect
to the instance briefly to obtain the size for listing.