job-list: don't respond after response error
Problem: several RPC handlers in the job-list module handle a
flux_respond() error by trying to send another response containing
an error, but that is likely to fail too, and is inconsistent
with the idiom used througout the flux-core code base.

Change logic to log the flux_respond() error and return without
attempting to respond again.