job-list: use flux_msglist for backlog
Problem: handling the journal backlog for pause/unpause with a
zlistx of json objects is slightly more complicated and less
intuitive than placing the response message in a flux_msglist.

Move to a flux_msglist.  As a bonus, items can be removed from
the list as it is iterated, rather than requiring a separate pass.