systemd: add housekeeping service unit
Problem: housekeeping scripts should be run in a systemd cgroup
for reliable termination, logging to the local systemd journal, and
debugging using well known systemd tools.

Add a systemd "oneshot" service unit for housekeeping, templated by
jobid and corresponding script that can be configured as an imp run target.
This is similar to what was proposed for prolog/epilog except, a default
30m runtime is enforced, and if the script fails, the node is drained.