shell/pmi: use f58plain for the kvsname
Problem: Flux uses the f58 jobid as the PMI "kvsname" and the ƒ
character encodes to 0xC6 0x92, but we now know that KVS values
are limited to 0x21 through 0x7E, so it may be best to avoid
using characters outside this range for the kvsname even though
that is currently unspecified.

Encode as "f58plain" instead.