libsdexec: add a channel input buffer
Problem: no buffer is created for input channels.

Add parameters to sdexec_channel_create_input() to support an
input buffer and credit callback, then create an input buffer
of the specified size, or a default value if zero.  The buffer
isn't used yet in this commit - that will be remedied later.

Update sdexec and the channel unit test to use the new function
signature, and rename the generic channel callback to indicate that
it is for output from channels, since there will soon be an input one.

Rename the internal 'writable' flag to 'is_input_channel', since
channels are unidirectional and adding buffering to input channels
could make this less obvious.