libsdexec: use the stdin channel buffer
Problem: the new channel stdin buffer is not used.

Have sdexec_channel_write() put data in the buffer and start the
io watcher.  Have the io watcher take data out of the buffer and
write it to the file descriptor.

Update unit test as reactor needs one more loop now to transfer
data to the file descriptor.

Fixes #6035