libsubprocess: use json object inside flux_cmd_t
Problem: since the flux_cmd_t object has to be represented
on the wire as a JSON object defined by RFC 42, it may be
simpler to maintain conformancy and be less surprising if it
is represented internally as JSON.

Convert internals to JSON.

Change two non-public functions from:
  zlist_t *cmd_channel_list (flux_cmd_t *cmd);
  zlist_t *cmd_msgchan_list (flux_cmd_t *cmd);
to
  json_t *cmd_get_channels (flux_cmd_t *cmd);
  json_t *cmd_get_msgchans (flux_cmd_t *cmd);

and update libsubprocess users.

In addition to alleviating the stated problems, the new
implementation is about 100 less lines of code.